src/az_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

jsonpickle
pandas
azure-functions
jsons
azure.identity
azure.mgmt.resourcehealth
pytest
azure-monitor-query